+++ dev/Common/WindowsAppRuntime.SelfContained.cpp
@@ -14,8 +14,6 @@
 #include "WindowsAppRuntime.VersionInfo.h"
 #include "WindowsAppRuntime.SelfContained.h"
 
-static std::wstring g_test_frameworkPackageFamilyName;
-
 STDAPI WindowsAppRuntime_IsSelfContained(
     BOOL* isSelfContained) noexcept try
 {
@@ -26,9 +24,7 @@ STDAPI WindowsAppRuntime_IsSelfContained(
     const PACKAGE_INFO* packageInfo{};
     wil::unique_cotaskmem_ptr<BYTE[]> buffer;
     RETURN_IF_FAILED(::AppModel::PackageGraph::GetCurrentPackageGraph(flags, packageInfoCount, packageInfo, buffer));
-    const auto frameworkPackageFamilyName{ !g_test_frameworkPackageFamilyName.empty() ?
-                                           g_test_frameworkPackageFamilyName :
-                                           ::WindowsAppRuntime::VersionInfo::GetPackageFamilyName() };
+    const auto frameworkPackageFamilyName{ ::WindowsAppRuntime::VersionInfo::GetPackageFamilyName() };
     for (uint32_t index=0; index < packageInfoCount; ++index)
     {
         if (CompareStringOrdinal(packageInfo[index].packageFamilyName, -1, frameworkPackageFamilyName.c_str(), -1, TRUE) == CSTR_EQUAL)
@@ -40,23 +36,6 @@ STDAPI WindowsAppRuntime_IsSelfContained(
 
     // Didn't find the Windows App SDK framework package in the package graph. We're self-contained!
     *isSelfContained = TRUE;
-    return S_OK;
-}
-CATCH_RETURN();
-
-STDAPI WindowsAppRuntime_SelfContainedTestInitialize(
-    PCWSTR frameworkPackageFamilyName) noexcept try
-{
-    if (!frameworkPackageFamilyName || (*frameworkPackageFamilyName == L'0'))
-    {
-        // Shutdown test support
-        g_test_frameworkPackageFamilyName.clear();
-    }
-    else
-    {
-        // Initialize test support
-        g_test_frameworkPackageFamilyName = frameworkPackageFamilyName;
-    }
     return S_OK;
 }
 CATCH_RETURN();
