+++ dev/WindowsAppRuntime_DLL/WindowsAppRuntime.def
@@ -20,6 +20,6 @@ EXPORTS
     GetSecurityDescriptorForAppContainerNames
 
     WindowsAppRuntime_IsSelfContained
-    WindowsAppRuntime_SelfContainedTestInitialize
 
     WindowsAppRuntime_VersionInfo_MSIX_Framework_PackageFamilyName_Get
+    WindowsAppRuntime_VersionInfo_TestInitialize
