+++ test/Common/Test_SelfContained.cpp
@@ -33,11 +33,11 @@ namespace Test::Common
         {
             ::TB::SetupBootstrap();
             auto cleanup = wil::scope_exit([&] {
-                ::WindowsAppRuntime::SelfContained::TestShutdown();
+                ::WindowsAppRuntime::VersionInfo::TestShutdown();
                 ::TB::CleanupBootstrap();
             });
             const auto c_doesNotExistPackageFamilyName{ L"Test.PackageFamilyName.DoesNotExist_1234567890abc" };
-            ::WindowsAppRuntime::SelfContained::TestInitialize(c_doesNotExistPackageFamilyName);
+            ::WindowsAppRuntime::VersionInfo::TestInitialize(c_doesNotExistPackageFamilyName);
 
             VERIFY_IS_TRUE(::WindowsAppRuntime::SelfContained::IsSelfContained());
         }
@@ -47,10 +47,10 @@ namespace Test::Common
             {
                 ::TB::SetupBootstrap();
                 auto cleanup = wil::scope_exit([&]{
-                    ::WindowsAppRuntime::SelfContained::TestShutdown();
+                    ::WindowsAppRuntime::VersionInfo::TestShutdown();
                     ::TB::CleanupBootstrap();
                 });
-                ::WindowsAppRuntime::SelfContained::TestInitialize(::TP::WindowsAppRuntimeFramework::c_PackageFamilyName);
+                WindowsAppRuntime::VersionInfo::TestInitialize(::TP::WindowsAppRuntimeFramework::c_PackageFamilyName);
 
                 VERIFY_IS_FALSE(::WindowsAppRuntime::SelfContained::IsSelfContained());
             }
@@ -60,16 +60,16 @@ namespace Test::Common
         {
             ::TB::SetupBootstrap();
             auto cleanup = wil::scope_exit([&] {
-                ::WindowsAppRuntime::SelfContained::TestShutdown();
+                ::WindowsAppRuntime::VersionInfo::TestShutdown();
                 ::TB::CleanupBootstrap();
             });
             const auto c_doesNotExistPackageFamilyName{ L"Test.PackageFamilyName.DoesNotExist_1234567890abc" };
-            ::WindowsAppRuntime::SelfContained::TestInitialize(c_doesNotExistPackageFamilyName);
+            WindowsAppRuntime::VersionInfo::TestInitialize(c_doesNotExistPackageFamilyName);
             VERIFY_IS_TRUE(::WindowsAppRuntime::SelfContained::IsSelfContained());
 
-            ::WindowsAppRuntime::SelfContained::TestInitialize(::TP::WindowsAppRuntimeFramework::c_PackageFamilyName);
+            WindowsAppRuntime::VersionInfo::TestInitialize(::TP::WindowsAppRuntimeFramework::c_PackageFamilyName);
             VERIFY_IS_FALSE(::WindowsAppRuntime::SelfContained::IsSelfContained());
-            ::WindowsAppRuntime::SelfContained::TestInitialize(c_doesNotExistPackageFamilyName);
+            WindowsAppRuntime::VersionInfo::TestInitialize(c_doesNotExistPackageFamilyName);
 
             VERIFY_IS_TRUE(::WindowsAppRuntime::SelfContained::IsSelfContained());
         }
