+++ test/Common/pch.h
@@ -36,6 +36,7 @@
 
 #include <Microsoft.Utf8.h>
 #include <WindowsAppRuntime.SelfContained.h>
+#include <WindowsAppRuntime.VersionInfo.h>
 
 #include <WindowsAppRuntime.Test.AppModel.h>
 #include <WindowsAppRuntime.Test.Package.h>
